Выделяет компонент 'Promo'.
