+++ src/components/main-page/main-page.jsx
@@ -1,14 +1,11 @@
 import React from "react";
 import {MovieCard} from "../movie-card/movie-card.jsx";
+import {Promo} from "../promo/promo.jsx";
 
 
 export const MainPage = (props) => {
   // eslint-disable-next-line react/prop-types
   const {promoMovie, movieTitles} = props;
-  // eslint-disable-next-line react/prop-types
-  const {title: promoMovieTitle, genre: promoMovieGenre, releaseDate, posterUrl: PromoMoviePosterUrl, backgroundUrl: promoBackgroundUrl} = promoMovie;
-  // eslint-disable-next-line react/prop-types
-  const promoMovieReleaseYear = releaseDate.getFullYear();
 
   return (
     <React.Fragment>
@@ -41,60 +38,7 @@ export const MainPage = (props) => {
         </svg>
       </div>
 
-      <section className="movie-card">
-        <div className="movie-card__bg">
-          <img src={promoBackgroundUrl} alt={promoMovieTitle} />
-        </div>
-
-        <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header movie-card__head">
-          <div className="logo">
-            <a className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="user-block">
-            <div className="user-block__avatar">
-              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-            </div>
-          </div>
-        </header>
-
-        <div className="movie-card__wrap">
-          <div className="movie-card__info">
-            <div className="movie-card__poster">
-              <img src={PromoMoviePosterUrl} alt={promoMovieTitle} width="218" height="327" />
-            </div>
-
-            <div className="movie-card__desc">
-              <h2 className="movie-card__title">{promoMovieTitle}</h2>
-              <p className="movie-card__meta">
-                <span className="movie-card__genre">{promoMovieGenre}</span>
-                <span className="movie-card__year">{promoMovieReleaseYear}</span>
-              </p>
-
-              <div className="movie-card__buttons">
-                <button className="btn btn--play movie-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list movie-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <span>My list</span>
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      <Promo promoMovie={promoMovie} />
 
       <div className="page-content">
         <section className="catalog">
